fix: 将商品 (products) 领域中涉及金额和货币的部分从 BigDecimal 类型更改为 long 类型, 并将货币换算等逻辑收敛到 CurrencyConfig 中

+++ international-shopping-domain/src/main/java/shopping/international/domain/model/vo/products/ProductPublicSnapshot.java
@@ -10,7 +10,6 @@ import shopping.international.domain.model.enums.products.ProductStatus;
 import shopping.international.domain.model.enums.products.SkuType;
 import shopping.international.types.utils.Verifiable;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.Collections;
 import java.util.List;
@@ -145,25 +144,25 @@ public class ProductPublicSnapshot implements Verifiable {
         @NotNull
         private String currency;
         /**
-         * 标价最小值
+         * 标价最小值（最小货币单位）
          */
         @Nullable
-        private BigDecimal listPriceMin;
+        private Long listPriceMin;
         /**
-         * 标价最大值
+         * 标价最大值（最小货币单位）
          */
         @Nullable
-        private BigDecimal listPriceMax;
+        private Long listPriceMax;
         /**
-         * 促销价最小值
+         * 促销价最小值（最小货币单位）
          */
         @Nullable
-        private BigDecimal salePriceMin;
+        private Long salePriceMin;
         /**
-         * 促销价最大值
+         * 促销价最大值（最小货币单位）
          */
         @Nullable
-        private BigDecimal salePriceMax;
+        private Long salePriceMax;
 
         /**
          * 校验币种字段
